+++ lib/routemaster/jobs/cache_and_sweep.rb
@@ -6,8 +6,8 @@ module Routemaster
     # Caches a URL using {Cache} and sweeps the dirty map if successful.
     # Busts the cache if the resource was deleted.
     class CacheAndSweep
-      def perform(url, options = {})
-        @source_peer = options[:source_peer]
+      def perform(url, client_options = {})
+        @client_options = client_options
         Dirty::Map.new.sweep_one(url) do
           begin
             cache.get(url)
@@ -22,18 +22,9 @@ module Routemaster
 
       def cache
         @cache ||= Routemaster::Cache.new(
-          client_options: client_options
+          client_options: @client_options
         )
       end
-
-      def client_options
-        if @source_peer
-          { source_peer: @source_peer }
-        else
-          {}
-        end
-      end
-
     end
   end
 end