+++ spec/routemaster/middleware/cache_spec.rb
@@ -20,7 +20,9 @@ RSpec.describe Routemaster::Middleware::Cache do
     let(:payload) { ['https://example.com/1'] }
 
     it 'queues a fetch job' do
-      expect(client).to receive(:enqueue).with('routemaster', Routemaster::Jobs::CacheAndSweep, 'https://example.com/1', nil)
+      expect(client).to receive(:enqueue).with(
+        'routemaster', Routemaster::Jobs::CacheAndSweep, 'https://example.com/1', nil
+        )
       perform
     end
   end